+++ package.json
@@ -38,6 +38,7 @@
     "react-dom": "^17.0.2",
     "react-indiana-drag-scroll": "^2.1.0",
     "react-markdown": "^8.0.0",
+    "react-router-dom": "^6.3.0",
     "remark-gfm": "^3.0.1",
     "stream-browserify": "^3.0.0",
     "util": "^0.12.4",
