+++ src/features/game/Goblins.tsx
@@ -1,0 +1,6 @@
+import React from "react";
+
+export const Goblins: React.FC = () => {
+  // Load data
+  return <span>TODO</span>;
+};
