+++ src/features/game/Session.tsx
@@ -1,43 +1,20 @@
-import React, { useRef, useEffect } from "react";
-import ScrollContainer from "react-indiana-drag-scroll";
+import React from "react";
 
-import background from "assets/land/background.png";
+import { BrowserRouter, Routes, Route, HashRouter } from "react-router-dom";
 
-import { Game } from "./Game";
-import { GameProvider } from "./GameProvider";
-import { ToastProvider } from "./toast/ToastQueueProvider";
-import mapMovement from "./lib/mapMovement";
-import { ExpansionInfo } from "./expansion/ExpansionInfo";
+import { Humans } from "./Humans";
+import { Goblins } from "./Goblins";
 
 export const Session: React.FC = () => {
-  // catching and passing scroll container to keyboard listeners
-  const container = useRef(null);
+  // Switch between humans and goblins using react router
 
-  useEffect(() => {
-    mapMovement.addListeners(container.current);
-    return () => {
-      mapMovement.removeListeners();
-    };
-  }, [container]);
-
-  // Load data
   return (
-    <GameProvider>
-      <ToastProvider>
-        <ScrollContainer
-          className="bg-green-background overflow-scroll relative w-full h-full"
-          innerRef={container}
-        >
-          <div
-            className="relative h-gameboard w-gameboard"
-            // TODO dynamic game board size based on tile dimensions
-          >
-            <img src={background} className="absolute inset-0 w-full h-full" />
-            <ExpansionInfo />
-            <Game />
-          </div>
-        </ScrollContainer>
-      </ToastProvider>
-    </GameProvider>
+    <HashRouter>
+      <Routes>
+        <Route path="/" element={<Humans />} />
+        <Route path="/humans" element={<Humans />} />
+        <Route path="/goblins" element={<Goblins />} />
+      </Routes>
+    </HashRouter>
   );
 };
